pyproject.toml
[project]
name = "fastapi-vue3-sample"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "sunzh231", email = "sunzh231@gmail.com"},
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "mcp-agent>=0.0.17",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "uvicorn main:app --reload"
test = "python ./backend/services/test.py"
